A Cool-looking bug.
Been working on adding a new feature to the shaders, and a bug makes it looks as
if the carges are casting some kind of shadows. Looks rather cool.

Add an initial README file
Add a line to gitignore for .gitattributes
Add a Result class that mimics the Result enum from Rust.
Move some code and classes from shader.gd into a new Shader_Utils.gd file.
Started to add a new feature to shaders - not finished & old functionality got bugged,